refact(file): rename guild-config.ts to guild-config-manager.ts
